Update the Privacy Policy and add a TOS (#75)
* started tos/privacy policy updates and implemented a guilds cache (why? see comment)
Messages in our message cache should be removed when a bot is kicked from a server, and by the time the guildremove event fires we no longer have access to the channels in those guilds. so we need this guild cache in order to know which channels we serve exactly and remove them from the cache.
Yes, we no longer serve those guilds anyways, but to air on the side of caution and avoid law intervention, we should say in our privacy policy that messages are no longer cached when the bot leaves your server. If nothing else its curtious to do that.

* finished update privacy policy

* finished privacy policy and tos

* remove the guild cache, because as it would turn out, LoadOrStore won't return data from channels it can't access, so the garbage collector handles this. what the fuck

* accidentally mentioned dforum in the first paragraph of the tos, should be dynamic basded on the name of the service

* accidentally mentioned dforum in the first paragraph of the tos, should be dynamic basded on the name of the service